# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Bussiness API",
        "title": "Bussiness API",
        "contact": {},
        "license": {
            "name": "Private"
        },
        "version": "1.0"
    },
    "host": "localhost:6060",
    "basePath": "/",
    "paths": {
        "/hello/{id}": {
            "get": {
                "description": "get string by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Show an account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.WorkflowTask"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.KV": {
            "type": "object",
            "additionalProperties": true
        },
        "types.WorkflowTask": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "conn": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "describe": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "params": {
                    "$ref": "#/definitions/types.KV"
                },
                "vars": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "access token",
            "type": "apiKey",
            "name": "X-AccessToken",
            "in": "header"
        }
    }
}